Add multi account handling

+++ run.py
@@ -19,6 +19,8 @@ class DiscordClient(discord.Client):
         self.current_date = datetime.strptime("2022-04-03", self.date_format)
         self.holder_role_ids = [x["role_id"] for x in settings.HOLDER_ROLES]
 
+        self.wsc_holders_db = self.pymongo_db.client.wsc_holders
+
         # get sabi (staking) contract object
         self.staking_contract_address = settings.STAKING_CONTRACT_ADDRESS
         self.count_staking_function = settings.COUNT_STAKING_FUNCTION
@@ -42,64 +44,51 @@ class DiscordClient(discord.Client):
         total_wsc_token_owned = len(staked_list) + unstaked_count
         return total_wsc_token_owned
 
-    def confirm_signed_message(self, member, signature):
-        return self.eth_client.get_address_of_signed_msg(member, signature)
+    def confirm_signed_message(self, member, signature, wallet_address):
+        try:
+            signed_wallet = self.eth_client.get_address_of_signed_msg(member, signature)
+        except:
+            signed_wallet = None
+        return signed_wallet == wallet_address
 
     def store_data_to_mongodb(self, member, holder_role, wallet_address, token_count):
         data = {
             "updated_at": datetime.utcnow(),
             "wallet_address": wallet_address,
-            "discord_id": member.id,
+            "discord_id": str(member.id),
             "token_count": token_count,
-            "holder_role": holder_role,
+            "holder_role": str(holder_role),
             }
-        wsc_holders_db = self.pymongo_db.client.wsc_holders
-        self.pymongo_db.insert_data_into_collection(wsc_holders_db.token_tracking, data)
+        self.pymongo_db.insert_data_into_collection(self.wsc_holders_db.token_tracking, data)
 
-    async def grant_role(self):
-        wallet_address = self.msg.split(" ")[-2]
-        signature = self.msg.split(" ")[-1]
-        member = self.message.author
+    def del_data_in_mongodb(self, discord_id):
+        search_query = {"discord_id": str(discord_id)}
+        self.pymongo_db.delete_in_collection(self.wsc_holders_db.token_tracking, search_query)
 
-        try:
-            confirmed_wallet_address = self.confirm_signed_message(str(member), signature)
-        except:
-            confirmed_wallet_address = None
-        
-        if wallet_address == confirmed_wallet_address:
-            token_count = self.get_wsc_token_count(wallet_address)
+    def get_current_holder_role(self, discord_id):
+        search_query = {"$match": {"discord_id": str(discord_id)}}
+        projection = {"holder_role": 1}
+        res = list(self.wsc_holders_db.token_tracking.aggregate([search_query, {"$project": projection}]))
+        return [int(r["holder_role"]) for r in res]
 
-            role_id = None
-            for holder in settings.HOLDER_ROLES:
-                if token_count >= holder["count"]:
-                    role_id = holder["role_id"]
-                    # end search when a role is assigned
-                    if role_id is not None:
-                        break
-                    else:
-                        continue
-            
-            self.store_data_to_mongodb(member, role_id, wallet_address, token_count)
+
+    def is_wallet_already_assigned(self, wallet_address):
+        search_query = {"$match": {"wallet_address": wallet_address}}
+        projection = {"discord_id": 1}
+        res = list(self.wsc_holders_db.token_tracking.aggregate([search_query, {"$project": projection}]))
+        return res[0]["discord_id"]
+    
+    async def remove_role(self, member, role_id, reason):
+        role = discord.utils.get(member.guild.roles, id=int(role_id))
+        await member.remove_roles(role, reason=reason)
+
+    async def grant_role(self, member, role_id):
             role = discord.utils.get(member.guild.roles, id=role_id)
             default_role = discord.utils.get(member.guild.roles, id=settings.DEFAULT_HOLDER_ROLE)
 
             await member.add_roles(role, reason="add specific nft holder count role")
             await member.add_roles(default_role, reason="add default nft holder count role")
-
-            # create embed message
-            embedMsg = discord.Embed(
-                    title="Your Sabis have been verified!",
-                    color=discord.Color.green(),
-                )
-
-            embedMsg.add_field(name="total WSC tokens found", value=token_count, inline=False)
-            await self.message.channel.send(embed=embedMsg)
             await self.message.channel.send(f"<@{member.id}>: `{role.name}` & `{default_role.name}` roles have been granted.")
-        else:
-            await self.message.channel.send(f'''
-            <@{member.id}>, Make sure you followed the correct format: `!verify [your wallet address] [sig]`
-Either you submitted an incorrect wallet address input or the signed message did not contain the right discord id message (e.g. test#1234). Please try again.
-            ''')
 
     async def remove_non_holders(self):
         await self.message.channel.send("Starting the check for non holders...")
@@ -164,12 +153,56 @@ Either you submitted an incorrect wallet address input or the signed message did
         #     self.current_date = message_date
         #     await self.update_holders()
         #     await self.remove_non_holders()
+
         if self.msg == "!verify":
             await self.message.channel.send(f'''Please follow this format `!verify [your wallet address] [sig]`''')
 
         elif self.msg.startswith("!verify"):
-            await self.grant_role()
-            await message.delete()  # delete message after granting the role
+            wallet_address = self.msg.split(" ")[-2]
+            signature = self.msg.split(" ")[-1]
+            member = self.message.author
+
+            confirmed_wallet_address = self.confirm_signed_message(str(member), signature, wallet_address)
+        
+            if confirmed_wallet_address:
+                discord_id_assigned = self.is_wallet_already_assigned(wallet_address)
+                if len(discord_id_assigned) > 0:
+                    await self.message.channel.send(f'''<@{member.id}>, this wallet has already been assigned to this discord account <@{discord_id_assigned}>. If you wish to use a different discord account, please reassign a new wallet to the old discord account first.''')
+                    await message.delete()  # delete message after granting the role
+                else:
+                    holder_roles = self.get_current_holder_role(member.id)
+                    for role_id in holder_roles:
+                        await self.remove_role(member, role_id, "removing existing roles for a new role.")
+                        self.del_data_in_mongodb(member.id)
+                    token_count = self.get_wsc_token_count(wallet_address)
+
+                    role_id = None
+                    for holder in settings.HOLDER_ROLES:
+                        if token_count >= holder["count"]:
+                            role_id = holder["role_id"]
+                            # end search when a role is assigned
+                            if role_id is not None:
+                                break
+                            else:
+                                continue
+                    
+                    self.store_data_to_mongodb(member, role_id, wallet_address, token_count)
+
+                    await self.grant_role(member, role_id)
+                    # create embed message
+                    embedMsg = discord.Embed(
+                            title="Your Sabis have been verified!",
+                            color=discord.Color.green(),
+                        )
+
+                    embedMsg.add_field(name="total WSC tokens found", value=token_count, inline=False)
+                    await self.message.channel.send(embed=embedMsg)
+                    await message.delete()  # delete message after granting the role
+            else:
+                await self.message.channel.send(f'''
+                <@{member.id}>, Make sure you followed the correct format: `!verify [your wallet address] [sig]`
+    Either you submitted an incorrect wallet address input or the signed message did not contain the right discord id message (e.g. test#1234). Please try again.
+                ''')
 
         elif self.msg.startswith("!connection"):
             await self.eth_connection()
